+++ src/main/java/racingGame/util/InputView.java
@@ -3,10 +3,12 @@ package racingGame.util;
 import java.util.Scanner;
 
 public class InputView {
-    public static String inputCar() {
+    public static String[] inputCar() {
         Scanner scanner = new Scanner(System.in);
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
-        return scanner.next();
+        String inputName = scanner.next();
+        String[] names = inputName.split(",");
+        return names;
     }
 
     public static int inputTry() {
